# app/http.go
// Trollr: A HTTP/JSON API for Troll
//
// Trollr is a simple wrapper around the amazing Troll: A dice roll language and calculator created by Torben Mogensen.
// The wrapper simply exposes and HTTP/JSON server that executes Troll, parses the results and returns it. The server
// has some built in rate-limiting and pooling to prevent abuse. I created this small server to support a Discord bot
// that I am working on.
//
//     Schemes: https
//     Host: trollr.live
//     BasePath: /api
//     Version: 0.1.0-alpha
//     License: MIT http://opensource.org/licenses/MIT
//     Contact: Ben Doerr <craftsman@bendoer.me> https://trollr.live
//
//     Consumes:
//     - application/json
//
//     Produces:
//     - application/json
//
//
// swagger:meta
package main

import (
	"context"
	"fmt"
	"net/http"
	"strconv"
	"time"

	"github.com/bendoerr/trollr/middleware"
	"github.com/didip/tollbooth"
	jsoniter "github.com/json-iterator/go"
	"github.com/json-iterator/go/extra"
	servertiming "github.com/mitchellh/go-server-timing"
	"go.uber.org/zap"
)

type API struct {
	server *http.Server
	mux    *http.ServeMux
	troll  *Troll
}

 func NewAPI(listen string, troll *Troll, logger *zap.Logger) *API {
	api := &API{
		troll: troll,
	}

	api.server = &http.Server{
		Addr:              listen,
		Handler:           nil,
		TLSConfig:         nil,
		ReadTimeout:       5 * time.Second,
		ReadHeaderTimeout: 5 * time.Second,
		WriteTimeout:      30 * time.Second,
		IdleTimeout:       60 * time.Second,
		MaxHeaderBytes:    0,
		TLSNextProto:      nil,
		ConnState:         nil,
		ErrorLog:          nil,
		BaseContext:       nil,
		ConnContext:       nil,
	}

	// Have jsoniter handle naming
	extra.SetNamingStrategy(extra.LowerCaseWithUnderscores)
	extra.RegisterFuzzyDecoders()

	api.mux = &http.ServeMux{}
	api.mux.Handle("/", http.HandlerFunc(api.DefaultNotFound))
	api.mux.Handle("/swagger.json", http.HandlerFunc(api.SwaggerJson))
	api.mux.Handle("/swagger", http.HandlerFunc(api.SwaggerRedirect))
	api.mux.Handle("/roll", http.HandlerFunc(api.Roll))
	api.mux.Handle("/calc", http.HandlerFunc(api.Calc))

	h := http.Handler(api.mux)
	h = middleware.GetMethodOnlyMiddleware(h)
	h = middleware.JsonContentTypeMiddleware(h)
	h = middleware.NoticeHeadersMiddleware(h)
	h = middleware.TimingMiddleware(h)
	h = middleware.LoggingMiddleware(h, logger)
	h = servertiming.Middleware(h, nil)
	h = middleware.PostponeWriteMiddleware(h)
	h = middleware.Recovery(h)
	h = tollbooth.LimitHandler(tollbooth.NewLimiter(1, nil), h)

	api.server.Handler = h

	return api
}

func (api *API) Start() {
	go func() {
		err := api.server.ListenAndServe()
		fmt.Println("server exited: ", err)
	}()
}

func (api *API) Stop() error {
	return api.server.Shutdown(context.Background())
}

func (api *API) DefaultNotFound(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotFound)
	_ = jsoniter.NewEncoder(w).Encode(RollsResult{
		Error: http.StatusText(http.StatusNotFound),
	})
}

func (api *API) SwaggerJson(w http.ResponseWriter, r *http.Request) {
	http.ServeFile(w, r, "static/swagger.json")
}

func (api *API) SwaggerRedirect(w http.ResponseWriter, r *http.Request) {
	http.Redirect(w, r,
		"https://app.swaggerhub.com/apis-docs/bendoerr/Trollr",
		http.StatusTemporaryRedirect)
}

// swagger:operation GET /roll API roll
//
// Roll Dice
//
// Given a roll definition this will delegate the roll to Troll and return the
// results structured as JSON.
//
// ---
// consumes:
// - application/json
// produces:
// - application/json
// parameters:
// - name: "d"
//   in: "query"
//   description: "The Troll roll definition"
//   type: "string"
//   required: true
// - name: "n"
//   in: "query"
//   description: "The number of times to repeat the roll"
//   type: "integer"
//   required: false
// responses:
//   '200':
//     description: "The results from rolling the dice"
//     schema:
//       "$ref": "#/definitions/RollsResult"
//   '400':
//     description: "The error will be populated in the result"
//     schema:
//       "$ref": "#/definitions/RollsResult"
func (api *API) Roll(w http.ResponseWriter, r *http.Request) {
	d := r.URL.Query().Get("d")
	n := r.URL.Query().Get("n")

	if len(d) < 1 {
		w.WriteHeader(http.StatusBadRequest)
		_ = jsoniter.NewEncoder(w).Encode(RollsResult{
			Error: "no roll definition given",
		})
		return
	}

	c := 1
	if len(n) > 0 {
		i, err := strconv.Atoi(n)
		if err == nil {
			c = i
		}
	}

	res := api.troll.MakeRolls(r.Context(), c, d)
	if res.Err != nil {
		w.WriteHeader(http.StatusBadRequest)
	}

	err := jsoniter.NewEncoder(w).Encode(res)
	if err != nil {
		fmt.Println(err)
	}
}

// swagger:operation GET /calc API calc
//
// Calculate the probabilities of dice roll.
//
// Given a roll definition this will delegate the roll to Troll and return the
// probabilities structured as JSON.
//
// ---
// consumes:
// - application/json
// produces:
// - application/json
// parameters:
// - name: "d"
//   in: "query"
//   description: "The Troll roll definition"
//   type: "string"
//   required: true
// - name: "c"
//   in: "query"
//   description: "What kind of cumulative probabilities you would like. One of 'ge' (default), 'gt', 'le', or 'lt'."
//   type: "string"
//   required: false
// responses:
//   '200':
//     description: "The probabilities of rolling the dice"
//     schema:
//       "$ref": "#/definitions/CalcResult"
//   '400':
//     description: "The error will be populated in the result"
//     schema:
//       "$ref": "#/definitions/CalcResult"
func (api *API) Calc(w http.ResponseWriter, r *http.Request) {
	d := r.URL.Query().Get("d")
	c := r.URL.Query().Get("c")

	if len(d) < 1 {
		w.WriteHeader(http.StatusBadRequest)
		_ = jsoniter.NewEncoder(w).Encode(RollsResult{
			Error: "no roll definition given",
		})
		return
	}

	res := api.troll.CalcRoll(r.Context(), d, c)
	if res.Err != nil {
		w.WriteHeader(http.StatusBadRequest)
	}

	err := jsoniter.NewEncoder(w).Encode(res)
	if err != nil {
		fmt.Println(err)
	}
}